Reuse ruby layer if required version is unchanged
This also means we need ruby in a cache dir so we can run bundler if the
ruby version is unchanged.

NOTE: using the cachedir for ruby also means that it will work for
subsequent buildpacks (multiple buildpacks). Realistically we should
probably also store the bundled gems in the cache (as well as a layer)
so that subsequent buildpacks can use them as well.